trueproof-service/requirements.txt
fastapi
uvicorn
tensorflow
pillow
numpy
resemblyzer
openai-whisper
torch
librosa
fpdf